+++ scripts/storage-emulator-integration/conformance/firebase.endpoints.test.ts
@@ -226,6 +226,29 @@ describe("Firebase Storage endpoint conformance tests", () => {
           .expect(200)
           .then((res) => new URL(res.header["x-goog-upload-url"]));
 
+        const uploadStatus = await supertest(firebaseHost)
+          .put(uploadURL.pathname + uploadURL.search)
+          .set({
+            "X-Goog-Upload-Protocol": "resumable",
+            "X-Goog-Upload-Command": "upload, finalize",
+            "X-Goog-Upload-Offset": 0,
+          })
+          .expect(403)
+          .then((res) => res.header["x-goog-upload-status"]);
+        expect(uploadStatus).to.equal("final");
+      });
+
+      it("should return 403 when resumable upload is unauthenticated and finalize is called again", async () => {
+        const testFileName = "disallowSize0";
+        const uploadURL = await supertest(firebaseHost)
+          .post(`/v0/b/${storageBucket}/o/${testFileName}?uploadType=resumable`)
+          .set({
+            "X-Goog-Upload-Protocol": "resumable",
+            "X-Goog-Upload-Command": "start",
+          })
+          .expect(200)
+          .then((res) => new URL(res.header["x-goog-upload-url"]));
+
         await supertest(firebaseHost)
           .put(uploadURL.pathname + uploadURL.search)
           .set({
@@ -234,6 +257,79 @@ describe("Firebase Storage endpoint conformance tests", () => {
             "X-Goog-Upload-Offset": 0,
           })
           .expect(403);
+        const uploadStatus = await supertest(firebaseHost)
+          .put(uploadURL.pathname + uploadURL.search)
+          .set({
+            "X-Goog-Upload-Protocol": "resumable",
+            "X-Goog-Upload-Command": "finalize",
+          })
+          .expect(403)
+          .then((res) => res.header["x-goog-upload-status"]);
+        expect(uploadStatus).to.equal("final");
+      });
+
+      it("should return 200 when resumable upload succeeds and finalize is called again", async () => {
+        const uploadURL = await supertest(firebaseHost)
+          .post(`/v0/b/${storageBucket}/o/${ENCODED_TEST_FILE_NAME}?uploadType=resumable`)
+          .set(authHeader)
+          .set({
+            "X-Goog-Upload-Protocol": "resumable",
+            "X-Goog-Upload-Command": "start",
+          })
+          .expect(200)
+          .then((res) => new URL(res.header["x-goog-upload-url"]));
+
+        await supertest(firebaseHost)
+          .put(uploadURL.pathname + uploadURL.search)
+          .set({
+            "X-Goog-Upload-Protocol": "resumable",
+            "X-Goog-Upload-Command": "upload, finalize",
+            "X-Goog-Upload-Offset": 0,
+          })
+          .expect(200);
+        await supertest(firebaseHost)
+          .put(uploadURL.pathname + uploadURL.search)
+          .set({
+            "X-Goog-Upload-Protocol": "resumable",
+            "X-Goog-Upload-Command": "finalize",
+          })
+          .expect(200);
+      });
+
+      it("should return 400 both times when finalize is called on cancelled upload", async () => {
+        const uploadURL = await supertest(firebaseHost)
+          .post(`/v0/b/${storageBucket}/o/${ENCODED_TEST_FILE_NAME}?uploadType=resumable`)
+          .set(authHeader)
+          .set({
+            "X-Goog-Upload-Protocol": "resumable",
+            "X-Goog-Upload-Command": "start",
+          })
+          .expect(200)
+          .then((res) => new URL(res.header["x-goog-upload-url"]));
+
+        await supertest(firebaseHost)
+          .put(uploadURL.pathname + uploadURL.search)
+          .set({
+            "X-Goog-Upload-Protocol": "resumable",
+            "X-Goog-Upload-Command": "cancel",
+            "X-Goog-Upload-Offset": 0,
+          })
+          .expect(200);
+        await supertest(firebaseHost)
+          .put(uploadURL.pathname + uploadURL.search)
+          .set({
+            "X-Goog-Upload-Protocol": "resumable",
+            "X-Goog-Upload-Command": "finalize",
+          })
+          .expect(400);
+
+        await supertest(firebaseHost)
+          .put(uploadURL.pathname + uploadURL.search)
+          .set({
+            "X-Goog-Upload-Protocol": "resumable",
+            "X-Goog-Upload-Command": "finalize",
+          })
+          .expect(400);
       });
     });