+++ src/emulator/storage/apis/firebase.ts
@@ -7,7 +7,12 @@ import { StorageEmulator } from "../index";
 import { EmulatorRegistry } from "../../registry";
 import { parseObjectUploadMultipartRequest } from "../multipart";
 import { NotFoundError, ForbiddenError } from "../errors";
-import { NotCancellableError, Upload, UploadNotActiveError } from "../upload";
+import {
+  NotCancellableError,
+  Upload,
+  UploadNotActiveError,
+  UploadPreviouslyFinalizedError,
+} from "../upload";
 import { reqBodyToBuffer } from "../../shared/request";
 import { ListObjectsResponse } from "../files";
 
@@ -206,6 +211,8 @@ export function createFirebaseEndpoints(emulator: StorageEmulator): Router {
         metadata = await storageLayer.uploadObject(upload);
       } catch (err) {
         if (err instanceof ForbiddenError) {
+          res.header("x-goog-upload-status", "final");
+          uploadService.setResponseCode(upload.id, 403);
           return res.status(403).json({
             error: {
               code: 403,
@@ -316,9 +323,14 @@ export function createFirebaseEndpoints(emulator: StorageEmulator): Router {
           upload = uploadService.finalizeResumableUpload(uploadId);
         } catch (err) {
           if (err instanceof NotFoundError) {
+            uploadService.setResponseCode(uploadId, 404);
             return res.sendStatus(404);
           } else if (err instanceof UploadNotActiveError) {
+            uploadService.setResponseCode(uploadId, 400);
             return res.sendStatus(400);
+          } else if (err instanceof UploadPreviouslyFinalizedError) {
+            res.header("x-goog-upload-status", "final");
+            return res.sendStatus(uploadService.getPreviousResponseCode(uploadId));
           }
           throw err;
         }