+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 - Adds support for media upload protocol to storage emulator (#4869)
 - Adds support to storage emualtor GCS API for /storage/v1/b/:bucketId/o/:objectId url pattern (#4869)
 - Removes extraneous default metadata fields from storage emulator (#4869)
+- Fixes an issue where a blocked upload from the android sdks crashes the storage emulator (#4860)